% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dl_wildfinder_data.R
\name{dl_wildfinder_data}
\alias{dl_wildfinder_data}
\title{Download Wildfinder dataset}
\usage{
dl_wildfinder_data(overwrite = FALSE)
}
\arguments{
\item{overwrite}{a \code{logical}. If \code{TRUE}, the file will be downloaded again
and the previous version will be replaced.}
}
\value{
No return value.
}
\description{
This function downloads the Wildfinder dataset (3 csv files) hosted on the
GitHub repository \url{https://github.com/rdatatoolbox/datarepo/}.
The files
won't be downloaded if already exists locally (except if \code{overwrite = TRUE}).
The files will be stored in
\verb{data/wildfinder/}. This folder will be created if required.
}
