% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dl_wwf_file.R
\name{dl_wwf_file}
\alias{dl_wwf_file}
\title{Download one file from Wildfinder dataset}
\usage{
dl_wwf_file(filename, overwrite, path, gurl)
}
\arguments{
\item{overwrite}{a \code{logical}. If \code{TRUE}, the file will be downloaded again
and the previous version will be replaced.}
}
\value{
No return value.
}
\description{
This function downloads a single file from wildfinder dataset)
The files
won't be downloaded if already exists locally (except if \code{overwrite = TRUE}).
The files will be stored in
\verb{data/wildfinder/}. This folder will be created if required.
}
